requirements.txt
###--CORE--##
chainlit
llama-index
langchain

###--TOOLS--###
trafilatura
wordcloud
matplotlib
pyperclip
PyPDF2
python-pptx
python-docx
pandas
openpyxl
tabulate